renamed notification tables
