+++ src/invites/components/AllInvitesList.tsx
@@ -2,7 +2,7 @@ import { useParam } from "@blitzjs/next"
 import { useQuery } from "@blitzjs/rpc"
 import getInvites from "../queries/getInvites"
 import Table from "src/core/components/Table"
-import { InviteTablePMColumns } from "../tables/columns/InviteTablePMColumns"
+import { InvitePMColumns } from "../tables/columns/InvitePMColumns"
 
 export const AllInvitesList = () => {
   const projectId = useParam("projectId", "number")
@@ -14,8 +14,7 @@ export const AllInvitesList = () => {
 
   return (
     <main className="flex flex-col mt-2 mx-auto w-full max-w-7xl">
-      {/* <h1 className="flex justify-center mb-2">All Contributors</h1> */}
-      <Table columns={InviteTablePMColumns} data={invites} addPagination={true} />
+      <Table columns={InvitePMColumns} data={invites} addPagination={true} />
     </main>
   )
 }
